batched semgrep
