+++ plugins/codemodder-plugin-semgrep/src/main/java/io/codemodder/providers/sarif/semgrep/SemgrepRule.java
@@ -1,0 +1,13 @@
+package io.codemodder.providers.sarif.semgrep;
+
+import java.nio.file.Path;
+import java.util.Objects;
+
+record SemgrepRule(SemgrepScan semgrepScan, String ruleId, Path yaml) {
+
+  SemgrepRule {
+    Objects.requireNonNull(semgrepScan);
+    Objects.requireNonNull(ruleId);
+    Objects.requireNonNull(yaml);
+  }
+}
