+++ plugins/codemodder-plugin-semgrep/src/main/java/io/codemodder/providers/sarif/semgrep/SemgrepRunner.java
@@ -11,12 +11,12 @@ public interface SemgrepRunner {
   /**
    * Execute semgrep.
    *
-   * @param yaml the file where the rule(s) are stored
+   * @param yamls the file(s) where the rule(s) are stored
    * @param codeDir the directory containing the code to be run on
    * @return the resulting SARIF
    */
   SarifSchema210 run(
-      Path yaml, Path codeDir, List<String> includePatterns, List<String> excludePatterns)
+      List<Path> yamls, Path codeDir, List<String> includePatterns, List<String> excludePatterns)
       throws IOException;
 
   static SemgrepRunner createDefault() {
